+++ test/random_test.dart
@@ -18,21 +18,18 @@ void main() {
       expect(randomBytes(100).length, 100);
     });
 
-    test('random buffer length = 0', () {
-      expect(randomBuffer(0), []);
-    });
-    test('random buffer length = 1', () {
-      expect(randomBuffer(1).lengthInBytes, 1);
-    });
-    test('random buffer length = 100', () {
-      expect(randomBuffer(100).lengthInBytes, 100);
-    });
-
-    test('random buffer length = 100', () {
+    test('fill random', () {
       var list = Uint32List(5);
       expect(list, equals([0, 0, 0, 0, 0]));
-      fillRandom(list.buffer);
-      expect(list, isNot(equals([0, 0, 0, 0, 0])));
+      int sum = 0;
+      for (int i = 0; i < 100; ++i) {
+        fillRandom(list.buffer, 5, 9);
+        var bytes = list.buffer.asUint8List();
+        sum += bytes.skip(5).take(9).any((n) => n > 0) ? 1 : 0;
+        expect(bytes.take(5).toList(), equals([0, 0, 0, 0, 0]));
+        expect(bytes.skip(5 + 9).toList(), equals([0, 0, 0, 0, 0, 0]));
+      }
+      expect(sum, greaterThan(0));
     });
   });
 }