+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 1.10.0
 
 - Adds support for `Poly1305` MAC generation: [#5](https://github.com/dipu-bd/hashlib/issues/5)
   - New class: `Poly1305`